test(llm): extract formatting tests to dedicated test modules
  - Move test_prompt_sdk_utilities.py → tests/llm/formatting/test_sdk_serialization.py
  - Create tests/llm/formatting/test_formatters.py with 4 extracted tests
  - Remove 4 formatting tests from test_prompt.py (verbose, raw, comparison tests)
  - Update imports to use public API without underscores
  - All formatting tests organized under tests/llm/formatting/
